+++ Database/populate.js
@@ -1,23 +1,21 @@
-const models = require("./models")
-
+const classes = require('./SequelizeClasses');
 
 async function create(){
+    await classes.sequelize.sync({ force: true });
+    
+    await classes.User.create({username: "james123", password:"1234", image: "https://www.w3schools.com/howto/img_avatar2.png"})
+    await classes.Project.create({name: "Choirs", description: "Keeps track of choirs", image: ""})
+    await classes.Column.create({name: "todo", Pid: 1})
+    await classes.Column.create({name: "inprogress", Pid: 1})
+    await classes.Column.create({name: "done", Pid: 1})
 
-    await models.sequelize.sync({ force: true });
-
-    await models.User.create({username: "james123", password:"1234", image: "https://www.w3schools.com/howto/img_avatar2.png"})
-    await models.Project.create({name: "Choirs", description: "Keeps track of choirs", image: ""})
-    await models.Column.create({name: "todo", project_id: 1})
-    await models.Column.create({name: "inprogress", project_id: 1})
-    await models.Column.create({name: "done", project_id: 1})
-
-    await models.Task.create({
-        name: "Do Laundry", 
-        description: "Complete the dirty laundry by putting it in the washing machine",
-        user_id: 1,
-        column_id: 1
+    await classes.Task.create({
+        name: "Do stuff",
+        description: "stuiff",
+        image: "",
+        Cid: 1,
+        Uid: 1
     })
 }
-
 
 create()
